refac: Move navigation out of main App.js file and into an infrastructure component
