+++ App.js
@@ -1,51 +1,15 @@
 import { StatusBar as ExpoStatusBar } from 'expo-status-bar';
 import AppLoading from 'expo-app-loading';
 import React from 'react';
-import { Text } from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { ThemeProvider } from 'styled-components/native';
 
-import { Ionicons } from '@expo/vector-icons';
 import { useFonts, Oswald_400Regular } from '@expo-google-fonts/oswald';
 import { Lato_400Regular } from '@expo-google-fonts/lato';
 
 import { theme } from './src/infrastructure/theme';
-import { SafeArea } from './src/components/SafeArea/SafeArea';
-import { RestaurantsScreen } from './src/features/restaurants/screens/RestaurantsScreen';
+import { Navigation } from './src/infrastructure/navigation';
 import { RestaurantsContextProvider } from './src/services/restaurants/RestaurantsContext';
 import { LocationContextProvider } from './src/services/location/LocationContext';
-
-const TAB_ICON = {
-  Restaurants: 'restaurant',
-  Map: 'map',
-  Settings: 'settings',
-};
-
-// ===== Mock Screen Components ===== //
-const Map = () => (
-  <SafeArea>
-    <Text>Maps!</Text>
-  </SafeArea>
-);
-const Settings = () => (
-  <SafeArea>
-    <Text>Settings!</Text>
-  </SafeArea>
-);
-// ===== Mock Screen Components ===== //
-
-const createScreenOptions = ({ route }) => {
-  const iconName = TAB_ICON[route.name];
-
-  return {
-    tabBarIcon: ({ size, color }) => (
-      <Ionicons name={iconName} size={size} color={color} />
-    ),
-  };
-};
-
-const Tab = createBottomTabNavigator();
 
 export default function App() {
   const [fontsLoaded] = useFonts({
@@ -61,22 +25,7 @@ export default function App() {
         <ThemeProvider theme={theme}>
           <LocationContextProvider>
             <RestaurantsContextProvider>
-              <NavigationContainer>
-                <Tab.Navigator
-                  screenOptions={createScreenOptions}
-                  tabBarOptions={{
-                    activeTintColor: 'tomato',
-                    inactiveTintColor: 'gray',
-                  }}
-                >
-                  <Tab.Screen
-                    name="Restaurants"
-                    component={RestaurantsScreen}
-                  />
-                  <Tab.Screen name="Map" component={Map} />
-                  <Tab.Screen name="Settings" component={Settings} />
-                </Tab.Navigator>
-              </NavigationContainer>
+              <Navigation />
             </RestaurantsContextProvider>
           </LocationContextProvider>
         </ThemeProvider>
