+++ src/infrastructure/navigation/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { AppNavigator } from './AppNavigator';
+
+export const Navigation = () => {
+  return <AppNavigator />;
+};
